Database relationships created

+++ src/model/agency.py
@@ -1,10 +1,10 @@
 from sqlalchemy import create_engine, MetaData, Table, Column, Integer, String, Float, select, ForeignKey
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker, relationship
+#from uuid import uuid1
 
-from .user import *
-#from .destination import *
-from .attraction import *
+#from .user import *
+from .attraction import * # import the Attraction and User class and create the tables in the database
 
 
 
@@ -19,7 +19,7 @@ class Agency(object):
         #self.destinations = destinations
         self.engine = create_engine('sqlite:///travel_app.db', echo=True)
         #self.connection = self.engine.connect() # connect to the database
-        self.loged_in_user = None # to keep track of the loged in user
+        self.loged_in_user_id = None # to keep track of the loged in user
 
     @staticmethod
     def get_instance():
@@ -30,10 +30,12 @@ class Agency(object):
 
         return Agency.singleton_instance
     
+
     def start_session(self):
         SessionClass = sessionmaker(bind=self.engine)
         session = SessionClass()
         return session
+        # closing the session is done in the functions that use it
         
         
     def create_user(self, name, type, password):
@@ -41,6 +43,7 @@ class Agency(object):
         
         try:
             user = User()
+            #user.id = uuid1()
             user.name = name
             user.type = type
             user.password = password
@@ -48,17 +51,19 @@ class Agency(object):
             session.add(user)
             session.commit() 
         except:
-            user = None
+            return user # return None if the user already exists
 
+        user = session.query(User).filter(User.name == name, User.password == password).first() # get the user from the database to access the id
+        self.loged_in_user_id = user.id # set the loged in user
         session.close() 
-        self.loged_in_users = user
         return user
     
     def get_user(self, name, password):
         session = self.start_session()
         user = session.query(User).filter(User.name == name, User.password == password).first() # get the user from the database .first() returns the first result or None
+        
+        self.loged_in_user_id = user.id
         session.close()
-        self.loged_in_users = user
         return user
     
     def get_destinations(self):
@@ -86,6 +91,8 @@ class Agency(object):
         return {1: "add attraction", 2: "remove attraction", 3: "update attraction", 4: "view attractions", 5: "logout"}
     
     def add_attraction(self, name, destination, attraction_type, price_range, description, contact, special_offer):
+
+        
         session = self.start_session()
         existing_attraction = session.query(Attraction).filter(Attraction.name == name, Attraction.destination == destination).first()
         if existing_attraction: # check if the attraction already exists
@@ -99,20 +106,55 @@ class Agency(object):
         attraction.description = description
         attraction.contact = contact
         attraction.special_offer = special_offer
+
+
+        attraction.provider_id = self.loged_in_user_id
+
         session.add(attraction)
         session.commit()
+        
+        # Load the currentlly loged in user
+        loged_in_user = session.query(User).get(self.loged_in_user_id)
+
+        # Refresh loged_in_user to make sure it has the latest data
+        #session.refresh(loged_in_user)
+
+        loged_in_user.attractions.append(attraction)
+
         session.close()
         return attraction
         
     def remove_attraction(self, name, destination):
         session = self.start_session()
         attraction = session.query(Attraction).filter(Attraction.name == name, Attraction.destination == destination).first()
-        if attraction:
+        if attraction and (attraction in self.loged_in_user_id.attractions): # check if the attraction exists and if it belongs to the provider
             session.delete(attraction)
             session.commit()
             session.close()
             return attraction
         session.close()
         return None
+    
+    def get_attraction(self, name, destination):
+        session = self.start_session()
+        attraction = session.query(Attraction).filter(Attraction.name == name, Attraction.destination == destination).first()
+        session.close()
+        if attraction:
+            return attraction
+        return "Attraction not found!"
+    
+    def get_attractions(self):
+        #session = self.start_session()
+        #attractions = session.query(Attraction).filter(Attraction in self.loged_in_user.attractions).all() # get all attractions from the provider
+        #session.close()
+        #attractions = sorted([attraction.name for attraction in attractions]) # get the names of the attractions
+        
+        ##### does this work? :
+        attractions = sorted([f"{attraction.name} in {attraction.destination}" for attraction in self.loged_in_user_id.attractions]) # get the names of the attractions
+        return attractions
+    
+    def get_attraction_details(attraction):
+        return f"Name: {attraction.name}\nDestination: {attraction.destination}\nType: {attraction.attraction_type}\nPrice range: {attraction.price_range}\nDescription: {attraction.description}\nContact: {attraction.contact}\nSpecial offer: {attraction.special_offer}\n" 
 
+
